refactor: mover resolvers to its own dir
