+++ src/mocks/handlers.ts
@@ -1,6 +1,6 @@
 import {http} from "msw";
-import {getUserResolver} from "./getUserResolver";
-import {sendCodeResolver} from "./sendCodeResolver";
+import {getUserResolver} from "./resolvers/getUserResolver";
+import {sendCodeResolver} from "./resolvers/sendCodeResolver";
 
 const backendUrl = import.meta.env.VITE_BACKEND_URL;
